Implemented the requested changes.
