+++ stk/molecular/molecules/__init__.py
@@ -6,8 +6,8 @@ from .molecule import *
 from .periodic import *
 from .polymer import *
 from .struct_unit import *
-from .macrocycles import *
-from .rotaxanes import *
+from .macrocycle import *
+from .rotaxane import *
 
 # Toggle the caching of molecules.
 OPTIONS['cache'] = True
