+++ stk/molecular/molecules/macrocycle.py
@@ -4,34 +4,12 @@ Defines classes that describe macrocycles.
 There is a family of classes dealing with macrocycles, a cyclic
 oligomer topology used to construct macrocycles.
 
-A base class :class:`MacrocycleBase` contains the methods used by all
-macrocycles, no matter whether loaded or constructed using ``stk``.
-A child class :class:`MacrocycleStructUnit` also inherits
-:class:`StructUnit` and is used to load macrocycles that were not
-constructed in ``stk``, while:class:`Macrocycle` inherits
-:class:`MacroMolecule` and is a result of construction within the
-``stk``. Either class can construct rotaxanes. The class
-:class:`MacrocycleBase` defines a couple of methods that are useful
-to generate properties of macrocycle.
-:meth:`MacrocycleBase.macro_atoms` returns the coordinates and indices
-of the atoms forming the largest ring in the macrocycle (used for
-threading). The method relies on the Smallest Set of Symmetric Rings
-and hence its result is not unique, but that should not cause any
-problem for most applications. A vector normal to the plane of the
-macrocycle is returned by :meth:`MacrocycleBase.macrocycle_plane'.
-
-One way to construct macrocycles is to use :meth:`Cyclic`, which
-behaves analogously to construction of linear polymers but the monomers
-are placed on a circumference of a large circle and the two terminal
-monomers are joined together to close the macrocycle. Otherwise it has
-attributes akin to :class:`Linear`, i.e. :attr:`Cyclic.repeating_unit`,
-:attr:`Cyclic.orientation`, and :attr:`Cyclic.n`.
-
 """
 
 from .macro_molecule import MacroMolecule
 from .struct_unit import StructUnit
 import rdkit.Chem.AllChem as rdkit
+import os
 
 
 class MacrocycleBase:
@@ -41,12 +19,33 @@ class MacrocycleBase:
     Macrocycles are molecules that contain a large cycle. A simple
     example is a polymer with two ends connected together. This base
     class allows for the macrocyles to be initialised as either
-    :class:`StructUnit` or :class:`MacroMolecule` but to be equally
+    :class:`.StructUnit` or :class:`.MacroMolecule` but to be equally
     identified as macrocycles.
 
     """
 
-    def macro_atoms(self, xyz=None, conformer=-1):
+    def cycle_atoms(self, conformer=-1):
+        """
+        Find the macrocyclic atoms in the molecule.
+
+        Ids of the atoms comprising the largest ring in the
+        macrocycle are returned. The method uses
+        :meth:`rdkit.GetSymSSSR()` to identify the Smallest Set of
+        Symmetric Rings, so the results are not unique. This should not
+        be a problem in most applications.
+
+        Returns
+        -------
+        :class:`.list` of :class:`.int`
+            Atom ids of the atoms comprising the largest ring.
+
+        """
+        ssr = rdkit.GetSymmSSSR(self.mol)
+        ring_atom_ids = list(max(ssr, key=len))
+
+        return ring_atom_ids
+
+    def cycle_coords(self, fpath=None, conformer=-1):
         """
         Find the macrocyclic atoms in the molecule.
 
@@ -59,30 +58,26 @@ class MacrocycleBase:
 
         Parameters
         ----------
-        xyz : :class:`str`
+        fpath : :class:`.str`
             A path where the xyz file should be saved. If None then
             no file is produced. InChKey is used as the filename.
 
         Returns
         -------
-        :class:`list` of :class:`list` of :class:`float`
+        :class:`.list` of :class:`.list` of :class:`.float`
             Coordinates of the atoms in the largest ring in the format
             [atomic_number, *xyz].
-
-        :class:`list` of :class:`int`
-            Atom ids of the atoms comprising the largest ring.
 
         """
         ssr = rdkit.GetSymmSSSR(self.mol)
         conf = self.mol.GetConformer(conformer)
-        ring_atom_ids = list(max(ssr, key=len))
         macrocycle = (self.mol.GetAtomWithIdx(i)
-                      for i in ring_atom_ids)
+                      for i in max(ssr, key=len))
         macro_coords = [[atom.GetAtomicNum(),
                          *conf.GetAtomPosition(atom.GetIdx())]
                         for atom in macrocycle]
 
-        if xyz is not None:
+        if fpath is not None:
             name = rdkit.MolToInchiKey(self.mol)
             xyz_file = f'{len(macro_coords)}\n\n'
 
@@ -90,18 +85,18 @@ class MacrocycleBase:
                 xyz_file += f'{anum} {coords[0]} {coords[1]} '
                 xyz_file += f'{coords[2]}\n'
 
-            if not os.path.exists(xyz):
-                os.makedirs(xyz)
+            if not os.path.exists(fpath):
+                os.makedirs(fpath)
 
-            with open(f'{xyz}/{name}.xyz', 'w') as f:
+            with open(f'{fpath}/{name}.xyz', 'w') as f:
                 f.write(xyz_file)
 
-        return macro_coords, ring_atom_ids
+        return macro_coords
 
 
 class MacrocycleStructUnit(MacrocycleBase, StructUnit):
     """
-    Used to represent macrocyles loaded as :class:`StructUnit`s.
+    Used to represent macrocyles loaded as :class:`.StructUnit`s.
 
     """
     pass
