+++ stk/molecular/molecules/rotaxane.py
@@ -1,0 +1,17 @@
+"""
+Defines classes that describe rotaxanes.
+
+A new :class:`.MacroMolecule` called :class:`.Rotaxane` is defined.
+This class stores constructed rotaxanes, normally originating from
+:class:`.NRotaxane` topology.
+
+"""
+
+from .macro_molecule import MacroMolecule
+
+class Rotaxane(MacroMolecule):
+    """
+    Used to represent rotaxanes.
+
+    """
+    pass
