+++ stk/molecular/topologies/__init__.py
@@ -3,4 +3,5 @@ from .base import *
 from .cof import *
 from .multimer import *
 from .polymer import *
-from .rotaxanes import *
+from .rotaxane import *
+from .macrocycle import *
